rmd-20676 Split TestCommonSetup from plugin tests
